fix(sports): align sport seed with registry + retire legacy positions module (Phase 2a)
Part of the baseball-deprecation effort. Two changes:

1. Migration `20260822120000_fix_sport_registry_parity.sql` (already applied):
   - Rename sport "Hockey" -> "Ice Hockey" so it matches the canonical registry
     key; DB name mismatch made position/metric lookups miss the registry.
   - Set has_position_list = true for Softball / Swimming / Track & Field, which
     have real multi-position registry lists but were wrongly flagged false.
     Rule: flag = (registry positions.length > 1); single-entry individual sports
     (Cross Country, Golf, Rowing, Wrestling) correctly stay false.

2. Retire the legacy baseball-only `utils/positions.ts` (BASEBALL_POSITIONS etc.).
   Its only live consumer (usePlayerDetailsForm) used two sport-agnostic helpers
   (shouldClearPositionOnSportChange, reconcilePositionOptions) — moved verbatim
   into utils/positions/canonical.ts. Deleted the flat module + its baseball-only
   tests; migrated the 8 sport-agnostic helper tests into canonicalPositions.spec.

Note: the seeded `positions` DB table (029) is orphaned — zero app queries; the
registry drives positions client-side. Left for a later drop migration.

type-check 0, lint 0, vitest 77/77 (positions/metrics suites).
